pyproject.toml
[tool.poetry]
name = "chess-video-analyzer"
version = "0.1.0"
description = "A tool for analyzing chess games from video recordings"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "chess_video_analyzer"}]

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.8.0"
python-chess = "^1.9.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mypy = "^1.3.0"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chess-video-analyzer = "chess_video_analyzer.cli.main:main"
